Finished Part 4

+++ Part 4/bloglist/controllers/blogs.js
@@ -1,15 +1,18 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', (request, response) => {
     Blog
       .find({})
+      .populate('user', {'username': 1, 'name': 1, 'id': 1})
       .then(blogs => {
         response.json(blogs)
       })
   })
   
-blogsRouter.post('/', (request, response) => {
+blogsRouter.post('/', async (request, response, next) => {
   if(request.body.likes === undefined){
     request.body.likes = 0
   }
@@ -17,19 +20,56 @@ blogsRouter.post('/', (request, response) => {
     response.status(400).end()
     return
   }
-  const blog = new Blog(request.body)
-  blog
-    .save()
-    .then(result => {
-      response.status(201).json(result)
-    })
-})
+  try{
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-blogsRouter.delete('/:id', (request, response) => {
-  Blog.findByIdAndRemove(request.params.id)
-  .then(() => {
-    response.status(204).end()
+    if(!token || !decodedToken.id){
+      return response.status(401).json({ error: 'token missing or invalid'})
+    }
+    const usr = await User.findById(decodedToken.id)
+    const blog = new Blog({
+      title: request.body.title,
+      author: request.body.author,
+      url: request.body.url,
+      likes: request.body.likes,
+      user: usr._id
+    })
+    blog
+      .save()
+      .then(result => {
+        usr.blogs = usr.blogs.concat(result._id)
+        usr.save()
+        response.status(201).json(result)
+      })
+    }
+    catch(exception){
+      next(exception)
+    }
   })
+
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try{
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if(!request.token || !decodedToken.id){
+      return response.status(401).json({ error: 'token missing or invalid'})
+    }
+    const blogToDel = await Blog.findById(request.params.id)
+    if(blogToDel.user.toString() !== decodedToken.id.toString()){
+      return response.status(401).json({ error: 'blog doesn\'t belong to the user'})
+    }
+    Blog.findByIdAndRemove(request.params.id)
+    .then(() => {
+      response.status(204).end()
+    })
+    const currUser = await User.findById(decodedToken.id)
+    currUser.blogs = currUser.blogs.filter(blog => {
+      blog.toString() !== request.params.id
+    })
+  }
+  catch(exception){
+    next(exception)
+  }
 })
 
 blogsRouter.put('/:id', (request, response) => {
